requirements.txt
torch>=1.9.0
torchvision>=0.10.0
opencv-python>=4.5.0
numpy>=1.21.0
matplotlib>=3.4.0
Pillow>=8.3.0
pyrealsense2>=2.50.0  # if RealSense camera is used